renamed treatment-layout and fixed some details on treatment card
