+++ src/interface/src/app/plan/create-scenarios/create-scenarios.component.scss
@@ -147,7 +147,7 @@
   align-items: center;
   flex-grow: 0;
   position: sticky;
-  z-index: 1200;
+  z-index: 500;
   border-top: 1px solid $color-soft-gray;
 }
