+++ src/interface/src/app/treatments/treatment-config/treatment-config.component.scss
