+++ src/interface/src/styleguide/treatment-card/treatment-card.component.scss
@@ -21,6 +21,7 @@
 
   &.disabled {
     box-shadow: none;
+    cursor: not-allowed;
 
     h4 {
       color: $color-text-gray;
